+++ core/src/main/java/com/goodjob/core/domain/member/service/MemberService.java
@@ -2,7 +2,6 @@ package com.goodjob.core.domain.member.service;
 
 
 import com.goodjob.core.domain.member.constant.ProviderType;
-import com.goodjob.core.domain.member.dto.request.EditRequestDto;
 import com.goodjob.core.domain.member.dto.request.JoinRequestDto;
 import com.goodjob.core.domain.member.entity.Member;
 import com.goodjob.core.domain.member.repository.MemberRepository;
@@ -112,47 +111,6 @@ public class MemberService {
         return socialJoin(providerType, username, "", email); // 최초 1회 실행
     }
 
-    public RsData<String> verifyPassword(String passwordToEdit, String memberPassword) {
-        if (!passwordEncoder.matches(passwordToEdit, memberPassword)) {
-            return RsData.of("F-1", "비밀번호가 일치하지않습니다.");
-        }
-
-        return RsData.of("S-1", "비밀번호가 일치합니다. 회원정보수정 페이지로 이동합니다.");
-    }
-
-    // 닉네임, 비밀번호 하나 이상 바꿔야 회원정보 수정 가능
-    @Transactional
-    public RsData update(Member member, EditRequestDto editRequestDto) {
-        String nickname = editRequestDto.getNickname().replaceAll("\\s+", "");
-        RsData nicknameVerificationRsData = verifyProvidedNickname(member.getNickname(), nickname);
-
-        if (nicknameVerificationRsData.isFail()) {
-            return nicknameVerificationRsData;
-        }
-
-        if (!member.isSocialMember()) {
-            boolean matches = passwordEncoder.matches(editRequestDto.getPassword(), member.getPassword());
-
-            // 기존 회원정보와 변화 없는 경우
-            if (member.getNickname().equals(nickname) && matches) {
-                return RsData.of("F-1", "수정된 정보가 없습니다!");
-            }
-
-            String password = passwordEncoder.encode(editRequestDto.getPassword());
-
-            member.setPassword(password);
-        } else {
-            if (member.getNickname().equals(nickname)) {
-                return RsData.of("F-1", "수정된 정보가 없습니다!");
-            }
-        }
-
-        member.setNickname(nickname);
-        memberRepository.save(member);
-
-        return RsData.of("S-1", "회원 정보가 수정되었습니다.");
-    }
-
     @Transactional
     public void delete(Long id) {
         memberRepository.deleteById(id);
@@ -186,16 +144,6 @@ public class MemberService {
         memberRepository.updateCoinForFreeMembers(MAX_COIN_COUNT);
     }
 
-    public EditRequestDto genEditRequestDtoWithTempPassword(Member member) {
-        String tempPassword = UUID.randomUUID().toString().replaceAll("-", "").substring(0, 15);
-        EditRequestDto editRequestDto = EditRequestDto.builder()
-                .nickname(member.getNickname())
-                .password(tempPassword)
-                .build();
-
-        return editRequestDto;
-    }
-
     public Optional<Member> findByNickname(String nickname) {
         return memberRepository.findByNickname(nickname);
     }
@@ -210,17 +158,6 @@ public class MemberService {
 
     public Optional<Member> findByEmail(String email) {
         return memberRepository.findByEmail(email);
-    }
-
-    private RsData verifyProvidedNickname(String originalNickname, String providedNickname) {
-        Optional<Member> opMember = findByNickname(providedNickname);
-
-        // 이미 존재하는 닉네임이면서 현재 닉네임과 수정하려는 닉네임이 같지 않은 경우
-        if (opMember.isPresent() && !originalNickname.equals(providedNickname)) {
-            return RsData.of("F-1", "이미 존재하는 닉네임 입니다.");
-        }
-
-        return RsData.of("S-1", "수정 가능한 닉네임 입니다.");
     }
 
     private Member genMember(String username, String password, String nickname, String email, ProviderType providerType) {
